Add release tooling: tool/repo.dart and release workflow
- Add tool/repo.dart (CommandRunner) with four commands:
  - check-versions: validates plugin.json, gemini-extension.json, and CHANGELOG.md are in sync
  - validate-manifests: checks all three manifest files for required fields
  - extract-changelog: prints the CHANGELOG.md section for a given version
  - generate-docs: regenerates README.md and docs/slipstream_doc.md (inlined from generate_docs.dart)
- Delete tool/validate_manifest.dart and tool/generate_docs.dart (superseded by repo.dart)
- Add .github/workflows/release.yml: on PRs, detects version bumps and labels/comments; on main, creates a GitHub release via gh release create
- Update ci.yml to use dart run tool/repo.dart validate-manifests
- Document release process in CONTRIBUTING.md
